# api/gofiber/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "keyword-search-backend API",
        "title": "keyword-search-backend Swagger API Documentation",
        "contact": {
            "name": "miniyus",
            "url": "https://miniyus.github.io",
            "email": "miniyu97@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/api/hosts/{id}/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get by host id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hosts"
                ],
                "summary": "get by host id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "host id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_keyword-search-backend_internal_search.ResponseByHost"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.ValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api_error.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "batch create search by host id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hosts"
                ],
                "summary": "batch create search by host id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "host id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "multi create search",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_keyword-search-backend_internal_search.MultiCreateSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_keyword-search-backend_internal_search.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.ValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/hosts/{id}/search/descriptions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get by host id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hosts"
                ],
                "summary": "get by host id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "host id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_keyword-search-backend_internal_search.DescriptionWithPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api_error.ValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/api_error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api_error.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "api_error.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "failed_fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_miniyus_keyword-search-backend_internal_search.CreateSearch": {
            "type": "object",
            "required": [
                "description",
                "host_id",
                "publish",
                "query",
                "query_key"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "host_id": {
                    "type": "integer"
                },
                "publish": {
                    "type": "boolean"
                },
                "query": {
                    "type": "string"
                },
                "query_key": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_keyword-search-backend_internal_search.Description": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "short_url": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_keyword-search-backend_internal_search.DescriptionWithPage": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_keyword-search-backend_internal_search.Description"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "github_com_miniyus_keyword-search-backend_internal_search.MultiCreateSearch": {
            "type": "object",
            "required": [
                "search"
            ],
            "properties": {
                "search": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_keyword-search-backend_internal_search.CreateSearch"
                    }
                }
            }
        },
        "github_com_miniyus_keyword-search-backend_internal_search.Response": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "publish": {
                    "type": "boolean"
                },
                "query": {
                    "type": "string"
                },
                "query_key": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_keyword-search-backend_internal_search.ResponseByHost": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_keyword-search-backend_internal_search.Response"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token type",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}